gh-95511: IDLE - fix Shell context menu copy-with-prompts bug (GH-95512)
If one selects whole lines, as the sidebar makes easy, do not
add an extra line.  Only move the end of a selection to the
beginning of the next line when not already at the beginning
of a line.  (Also improve the surrounding code.)
(cherry picked from commit f5d7b12d50b950027d9f465554d02a002b25caae)

Co-authored-by: Terry Jan Reedy <tjreedy@udel.edu>